requirements.txt
fastapi
uvicorn
dnspython
requests
python-whois
